adding more responsiveness.
The homepage can be scrolled in both x and y axis  or  we can say leftward or  downward, which does not given very good user experience.

In order to improve the user experience i have basically used flex wrap, rem, em like properties in order to make it a little better.